+++ Makefile
@@ -1,7 +1,7 @@
 build:
 	@(rm -rf build || true)
-	./node_modules/.bin/lys test/test.lys --test
-	./node_modules/.bin/lys src/main.lys --wast
+	# ./node_modules/.bin/lys test/test.lys --test
+	./node_modules/.bin/lys src/main.lys --wast --lib node_modules/lys/dist/utils/libs/env.js
 	@cd src; ../node_modules/.bin/ncc build index.ts -o ../dist
 	@./node_modules/.bin/mocha -r ts-node/register test/\*\*/\*.spec.ts
