+++ src/parser/parser.lys
@@ -52,7 +52,7 @@ fun parse(ruleName: string, parser: Parser, level: i32): AstNode | Nil = {
     }
     else -> {
       support::env::printf("MISSING RULE " ++ ruleName ++ "!")
-      Nil
+      Rule0
     }
   }
 }
@@ -63,13 +63,19 @@ fun eatUnexpectedToken(lexer: Lexer): AstNode = {
   UnexpectedToken(token, tokenValue)
 }
 
+val TEST = false
+
 fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNode | Nil = {
   var backtrack = parser.lexer.pos
 
+  if (level > 1000) {
+  support::env::printf(repeat("-", level - 1000) ++ ruleName)
+  }
+
   val result = match rule {
     case is Terminal(tokenType) -> {
-      support::env::printf(repeat("| ", level) ++ "| Terminal(" ++ TokenType.toString(tokenType) ++ ")")
-      support::env::printf(repeat("| ", level) ++ "|   from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Terminal(" ++ TokenType.toString(tokenType) ++ ")")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "|   from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
       val lexer = parser.lexer
       val token = lexer.eat()
       if (token.tokenType == tokenType)
@@ -78,8 +84,8 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
         Nil
     }
     case is StrictTerminal(tokenType, value) -> {
-      support::env::printf(repeat("| ", level) ++ "| StrictTerminal(" ++ TokenType.toString(tokenType) ++ ")")
-      support::env::printf(repeat("| ", level) ++ "|   from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| StrictTerminal(" ++ TokenType.toString(tokenType) ++ ")")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "|   from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
       val lexer = parser.lexer
       val token = lexer.eat()
       val tokenValue = token.textIn(lexer.source)
@@ -89,14 +95,14 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
         Nil
     }
     case is Or(lhs, rhs) -> {
-      support::env::printf(repeat("| ", level) ++ "| Or")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Or")
       match parse(lhs, ruleName, parser, level + 1) {
         case is Nil -> parse(rhs, ruleName, parser, level + 1)
         case ast is AstNode -> ast
       }
     }
     case is Cons(head, tail) -> {
-      support::env::printf(repeat("| ", level) ++ "| Cons")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Cons")
       match parse(head, ruleName, parser, level + 1) {
         case is Nil -> Nil
         case astHead is AstNode -> match parse(tail, ruleName, parser, level + 1) {
@@ -107,7 +113,7 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
     }
     case is Cut(head, tail) -> {
       // similar to Cons but with cut semantics
-      support::env::printf(repeat("| ", level) ++ "| Cut")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Cut")
       match parse(head, ruleName, parser, level + 1) {
         case is Nil -> Nil
         case astHead is AstNode -> {
@@ -127,25 +133,25 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
       }
     }
     case is Discard(rule) -> {
-      support::env::printf(repeat("| ", level) ++ "| Discard")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Discard")
       match parse(rule, ruleName, parser, level + 1) {
         case is Nil -> Nil
         case ast is AstNode -> Rule0
       }
     }
     case is Optional(rule) -> {
-      support::env::printf(repeat("| ", level) ++ "| Optional")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| Optional")
       match parse(rule, ruleName, parser, level + 1) {
         case is Nil -> Rule0
         case ast is AstNode -> ast
       }
     }
     case is NonTerminal(name) -> {
-      support::env::printf(repeat("| ", level) ++ "| NonTerminal("++name++")")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| NonTerminal("++name++")")
       parse(name, parser, level + 1)
     }
     case is OneOrMore(rule) -> {
-      support::env::printf(repeat("| ", level) ++ "| OneOrMore")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| OneOrMore")
       var ret: AstNode | Nil = Nil
 
       loop {
@@ -164,7 +170,7 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
       ret
     }
     case is ZeroOrMore(rule) -> {
-      support::env::printf(repeat("| ", level) ++ "| ZeroOrMore")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| ZeroOrMore")
       var ret: AstNode = Rule0
 
       loop {
@@ -180,13 +186,23 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
       ret
     }
     case is LookAhead(rule) -> {
-      support::env::printf(repeat("| ", level) ++ "| LookAhead")
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| LookAhead")
       val ret = parse(rule, ruleName, parser, level + 1)
       // rewind
       parser.lexer.pos = backtrack
       match ret {
         case is Nil -> Nil
         else -> Rule0
+      }
+    }
+    case is NegativeLookAhead(rule) -> {
+      if (TEST) support::env::printf(repeat("| ", level) ++ "| NegativeLookAhead")
+      val ret = parse(rule, ruleName, parser, level + 1)
+      // rewind
+      parser.lexer.pos = backtrack
+      match ret {
+        case is Nil -> Rule0
+        else -> Nil
       }
     }
     case is Fail(message) -> {
@@ -199,12 +215,12 @@ fun parse(rule: ParserRule, ruleName: string, parser: Parser, level: i32): AstNo
 
   if (result is Nil) {
     if (parser.lexer.pos != backtrack){
-      support::env::printf(repeat("| ", level) ++ "|<- backtracking from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
+      if (TEST) support::env::printf(repeat("| ", level) ++ "|<- backtracking from:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
       parser.lexer.pos = backtrack
-      support::env::printf(repeat("| ", level) ++ "|<- backtracking to:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
+      if (TEST) support::env::printf(repeat("| ", level) ++ "|<- backtracking to:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
     }
   } else {
-    support::env::printf(repeat("| ", level) ++ "| push. new string:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
+    if (TEST) support::env::printf(repeat("| ", level) ++ "| push. new string:" ++ parser.lexer.source.substring(parser.lexer.pos as i32))
   }
 
   result
