+++ test/fbt.ts
@@ -1,40 +1,25 @@
 import { readFileSync, existsSync, writeFileSync } from "fs";
 import * as expect from "expect";
 import * as glob from "glob";
-import tokenizer from "../dist";
-import { resolve } from "path";
 
-const writeToFile = process.env.UPDATE_AST;
 
-export function folderBasedTest(grep: string) {
-  const unknowns = new Set<string>();
+export const WRITE_TO_FILE = process.env.UPDATE_AST;
+
+
+export function folderBasedTest(grep: string, fn: (source: string) => Promise<string>, extension: string) {
 
   function testFile(fileName: string) {
     it(fileName, async () => {
-      let instance = await tokenizer();
       const content = readFileSync(fileName).toString();
-      instance.startLexer(content);
-
-      const tokens: string[] = [];
-      let token: string;
-
-      do {
-        token = instance.eat();
-        if (token.startsWith("Unknown(")) {
-          unknowns.add(token.replace("Unknown(", '').replace(/\)$/, ''));
-        }
-        tokens.push(JSON.stringify(token));
-      } while (token !== "EndOfFile");
-
-      const result = tokens.join("\n");
+      const result = await fn(content)
 
       if (result !== null) {
-        const compareToFileName = fileName + ".result";
+        const compareToFileName = fileName + extension;
         const compareFileExists = existsSync(compareToFileName);
         const compareTo = compareFileExists
           ? readFileSync(compareToFileName).toString()
           : "";
-        if (writeToFile || !compareFileExists) {
+        if (WRITE_TO_FILE || !compareFileExists) {
           writeFileSync(compareToFileName, result);
         }
         expect(compareTo.trim().length > 0 || !compareFileExists).toEqual(true);
@@ -47,26 +32,4 @@ export function folderBasedTest(grep: string) {
     glob.sync(grep).map(testFile);
   });
 
-  describe("Compares unknowns", () => {
-    it("compares against golden file", () => {
-      const result = Array.from(unknowns)
-        .sort()
-        .join("\n");
-
-      if (result !== null) {
-        const compareToFileName = resolve(
-          __dirname,
-          "fixtures/unknowns.result"
-        );
-        const compareFileExists = existsSync(compareToFileName);
-        const compareTo = compareFileExists
-          ? readFileSync(compareToFileName).toString()
-          : "";
-        if (writeToFile || !compareFileExists) {
-          writeFileSync(compareToFileName, result);
-        }
-        expect(result.trim()).toEqual(compareTo.trim());
-      }
-    });
-  });
 }
