--- test/fileBasedTest.spec.ts
@@ -1,4 +1,0 @@
-import { folderBasedTest } from "./fbt";
-import { resolve } from "path";
-
-folderBasedTest(resolve(__dirname, "./fixtures/") + "/**/*.lys");
