+++ test/fixtures/core.lys.ast
@@ -1,0 +1,7 @@
+|-Document
+  |-Directive
+    |-TypeDirective
+      |-TYPE_KEYWORD
+        |-Identifier=type
+      |-NameIdentifier
+        |-Identifier=void
