+++ test/fixtures/stdlib/system/DataView.lys.ast
@@ -1,0 +1,136 @@
+|-Document
+  |-Directive
+    |-StructDirective
+      |-STRUCT_KEYWORD
+        |-Identifier=struct
+      |-StructDeclaration
+        |-NameIdentifier
+          |-Identifier=DataView
+        |-FunctionParamsList
+          |-ParameterList
+            |-Parameter
+              |-NameIdentifier
+                |-Identifier=buffer
+              |-OfType
+                |-Operator=:
+                |-Type
+                  |-UnionType
+                    |-IntersectionType
+                      |-AtomType
+                        |-Reference
+                          |-QName
+                            |-NameIdentifier
+                              |-Identifier=bytes
+  |-Directive
+    |-ImplDirective
+      |-IMPL_KEYWORD
+        |-Identifier=impl
+      |-Reference
+        |-QName
+          |-NameIdentifier
+            |-Identifier=DataView
+      |-NamespaceElementList
+        |-CurlyBracesOpen={
+        |-<UNEXPECTED TOKEN>=fun
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=[
+        |-<UNEXPECTED TOKEN>=]
+        |-<UNEXPECTED TOKEN>=(
+        |-<UNEXPECTED TOKEN>=self
+        |-<UNEXPECTED TOKEN>=:
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=DataView
+        |-<UNEXPECTED TOKEN>=,
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=at
+        |-<UNEXPECTED TOKEN>=:
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=u32
+        |-<UNEXPECTED TOKEN>=)
+        |-<UNEXPECTED TOKEN>=:
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=u8
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>==
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>={
+        |-<UNEXPECTED TOKEN>=
+
+        |-<UNEXPECTED TOKEN>=    
+        |-<UNEXPECTED TOKEN>=if
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=(
+        |-<UNEXPECTED TOKEN>=at
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=>
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>=self
+        |-<UNEXPECTED TOKEN>=.
+        |-<UNEXPECTED TOKEN>=buffer
+        |-<UNEXPECTED TOKEN>=.
+        |-<UNEXPECTED TOKEN>=length
+        |-<UNEXPECTED TOKEN>=)
+        |-<UNEXPECTED TOKEN>= 
+        |-<UNEXPECTED TOKEN>={
+        |-<UNEXPECTED TOKEN>=
+
+        |-<UNEXPECTED TOKEN>=      
+        |-<UNEXPECTED TOKEN>=panic
+        |-<UNEXPECTED TOKEN>=(
+        |-<UNEXPECTED TOKEN>=)
+        |-<UNEXPECTED TOKEN>=
+
+        |-<UNEXPECTED TOKEN>=    
+        |-CurlyBracesClose=}
+  |-<UNEXPECTED TOKEN>=if
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=at
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=<
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=0
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=as
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=u32
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=      
+  |-<UNEXPECTED TOKEN>=panic
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=    
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=    
+  |-<UNEXPECTED TOKEN>=u8
+  |-<UNEXPECTED TOKEN>=.
+  |-<UNEXPECTED TOKEN>=load
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=self
+  |-<UNEXPECTED TOKEN>=.
+  |-<UNEXPECTED TOKEN>=buffer
+  |-<UNEXPECTED TOKEN>=.
+  |-<UNEXPECTED TOKEN>=ptr
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=as
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=ref
+  |-<UNEXPECTED TOKEN>=,
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=at
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=}
