+++ test/fixtures/stdlib/system/core.lys.ast
@@ -1,0 +1,171 @@
+|-Document
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=native
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=boolean
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=u8
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=u16
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=u32
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=u64
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=i16
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=i32
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=i64
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=f32
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=f64
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=bytes
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=string
+  |-EndOfFile=
