+++ test/fixtures/stdlib/system/core/rtti.lys.ast
@@ -1,0 +1,123 @@
+|-Document
+  |-Directive
+    |-ImportDirective
+      |-IMPORT_KEYWORD
+        |-Identifier=import
+      |-QName
+        |-NameIdentifier
+          |-Identifier=system
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=core
+        |-Operator=::
+        |-NameIdentifier
+          |-Identifier=native
+  |-Directive
+    |-FunctionDirective
+      |-Decorators
+        <Rule0>
+      |-FunDeclaration
+        |-FunctionName
+          |-NameIdentifier
+            |-Identifier=retain
+        |-FunctionParamsList
+          |-ParameterList
+            |-Parameter
+              |-NameIdentifier
+                |-Identifier=reference
+              |-OfType
+                |-Operator=:
+                |-Type
+                  |-UnionType
+                    |-IntersectionType
+                      |-AtomType
+                        |-Reference
+                          |-QName
+                            |-NameIdentifier
+                              |-Identifier=ref
+        |-OfType
+          |-Operator=:
+          |-Type
+            |-UnionType
+              |-IntersectionType
+                |-AtomType
+                  |-Reference
+                    |-QName
+                      |-NameIdentifier
+                        |-Identifier=void
+        |-FunAssignExpression
+          |-Operator==
+          |-Expression
+            |-AssignExpression
+              |-OrExpression
+                |-AndExpression
+                  |-BitOrExpression
+                    |-BitXorExpression
+                      |-BitAndExpression
+                        |-EqExpression
+                          |-RelExpression
+                            |-ShiftExpression
+                              |-AddExpression
+                                |-MulExpression
+                                  |-IsExpression
+                                    |-AsExpression
+                                      |-UnaryExpression
+                                        |-AtomicExpression
+                                          |-Value
+                                            |-CodeBlock
+                                              <Rule0>
+  |-Directive
+    |-FunctionDirective
+      |-Decorators
+        <Rule0>
+      |-FunDeclaration
+        |-FunctionName
+          |-NameIdentifier
+            |-Identifier=release
+        |-FunctionParamsList
+          |-ParameterList
+            |-Parameter
+              |-NameIdentifier
+                |-Identifier=reference
+              |-OfType
+                |-Operator=:
+                |-Type
+                  |-UnionType
+                    |-IntersectionType
+                      |-AtomType
+                        |-Reference
+                          |-QName
+                            |-NameIdentifier
+                              |-Identifier=ref
+        |-OfType
+          |-Operator=:
+          |-Type
+            |-UnionType
+              |-IntersectionType
+                |-AtomType
+                  |-Reference
+                    |-QName
+                      |-NameIdentifier
+                        |-Identifier=void
+        |-FunAssignExpression
+          |-Operator==
+          |-Expression
+            |-AssignExpression
+              |-OrExpression
+                |-AndExpression
+                  |-BitOrExpression
+                    |-BitXorExpression
+                      |-BitAndExpression
+                        |-EqExpression
+                          |-RelExpression
+                            |-ShiftExpression
+                              |-AddExpression
+                                |-MulExpression
+                                  |-IsExpression
+                                    |-AsExpression
+                                      |-UnaryExpression
+                                        |-AtomicExpression
+                                          |-Value
+                                            |-CodeBlock
+                                              <Rule0>
+  |-EndOfFile=
