+++ test/fixtures/stdlib/system/gc.lys.ast
@@ -1,0 +1,138 @@
+|-Document
+  |-Directive
+    |-TypeDirective
+      |-TYPE_KEYWORD
+        |-Identifier=type
+      |-NameIdentifier
+        |-Identifier=Object
+  |-<UNEXPECTED TOKEN>==
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=%
+  |-<UNEXPECTED TOKEN>=struct
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+
+  |-<UNEXPECTED TOKEN>=impl
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=Object
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=fun
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=apply
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=size
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=i32
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=Object
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>==
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=%
+  |-<UNEXPECTED TOKEN>=wasm
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=    
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=call
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=$system
+  |-<UNEXPECTED TOKEN>=::
+  |-<UNEXPECTED TOKEN>=core
+  |-<UNEXPECTED TOKEN>=::
+  |-<UNEXPECTED TOKEN>=memory
+  |-<UNEXPECTED TOKEN>=::
+  |-<UNEXPECTED TOKEN>=calloc
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=i32
+  |-<UNEXPECTED TOKEN>=.
+  |-<UNEXPECTED TOKEN>=add
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=i32
+  |-<UNEXPECTED TOKEN>=.
+  |-<UNEXPECTED TOKEN>=const
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=16
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=get_local
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=$size
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=fun
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=link
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=parent
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=ref
+  |-<UNEXPECTED TOKEN>=,
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=child
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=ref
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=void
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>==
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=/** Unlinks this object from its list. */
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=  
+  |-<UNEXPECTED TOKEN>=fun
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=unlink
+  |-<UNEXPECTED TOKEN>=(
+  |-<UNEXPECTED TOKEN>=self
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=ref
+  |-<UNEXPECTED TOKEN>=)
+  |-<UNEXPECTED TOKEN>=:
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>=void
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>==
+  |-<UNEXPECTED TOKEN>= 
+  |-<UNEXPECTED TOKEN>={
+  |-<UNEXPECTED TOKEN>=}
+  |-<UNEXPECTED TOKEN>=
+
+  |-<UNEXPECTED TOKEN>=}
