+++ test/fixtures/stdlib/system/random.lys.ast
@@ -1,0 +1,99 @@
+|-Document
+  |-Directive
+    |-FunctionDirective
+      |-Decorators
+        <Rule0>
+      |-FunDeclaration
+        |-FunctionName
+          |-NameIdentifier
+            |-Identifier=nextInt
+        |-FunctionParamsList
+          |-ParameterList
+            <Rule0>
+        |-OfType
+          |-Operator=:
+          |-Type
+            |-UnionType
+              |-IntersectionType
+                |-AtomType
+                  |-Reference
+                    |-QName
+                      |-NameIdentifier
+                        |-Identifier=i32
+        |-FunAssignExpression
+          |-Operator==
+          |-Expression
+            |-AssignExpression
+              |-OrExpression
+                |-AndExpression
+                  |-BitOrExpression
+                    |-BitXorExpression
+                      |-BitAndExpression
+                        |-EqExpression
+                          |-RelExpression
+                            |-ShiftExpression
+                              |-AddExpression
+                                |-MulExpression
+                                  |-IsExpression
+                                    |-AsExpression
+                                      |-UnaryExpression
+                                        |-AtomicExpression
+                                          |-Value
+                                            |-Literal
+                                              |-PostfixNumber
+                                                |-NumberLiteral=0
+  |-Directive
+    |-FunctionDirective
+      |-Decorators
+        <Rule0>
+      |-FunDeclaration
+        |-FunctionName
+          |-NameIdentifier
+            |-Identifier=nextFloat
+        |-FunctionParamsList
+          |-ParameterList
+            <Rule0>
+        |-OfType
+          |-Operator=:
+          |-Type
+            |-UnionType
+              |-IntersectionType
+                |-AtomType
+                  |-Reference
+                    |-QName
+                      |-NameIdentifier
+                        |-Identifier=f32
+        |-FunAssignExpression
+          |-Operator==
+          |-Expression
+            |-AssignExpression
+              |-OrExpression
+                |-AndExpression
+                  |-BitOrExpression
+                    |-BitXorExpression
+                      |-BitAndExpression
+                        |-EqExpression
+                          |-RelExpression
+                            |-ShiftExpression
+                              |-AddExpression
+                                |-MulExpression
+                                  |-IsExpression
+                                    |-AsExpression
+                                      |-UnaryExpression
+                                        |-AtomicExpression
+                                          |-Value
+                                            |-Literal
+                                              |-PostfixNumber
+                                                |-NumberLiteral=0
+                                          |-MemberExpression
+                                            |-BinMemberOperator
+                                              |-Operator=.
+                                            |-<UNEXPECTED TOKEN>=0
+                                            |-<UNEXPECTED TOKEN>=
+
+
+                                            |-<UNEXPECTED TOKEN>=// fun nextLong(): i64 = 0
+
+                                            |-<UNEXPECTED TOKEN>=// fun nextDouble(): f64 = 0.0
+
+  |-EndOfFile=
